uploaded revision 5
Created a "describe the purpose" section, cleaned up the data understanding section, created a "verify data quality" section, and completed a simple statistics section.  At this point the only four sections that remain to be completed are the attribute relationships, relationships between the attributes and the predicted class, other features that can be added, and exceptional work.