Added rev 2
Completed the deployment section, mostly completed both data preparation sections.  This work is assuming that we will be using the income dataset and doing one classification on income and one classification on marital status.  If we decide to do a different dataset or different classification tasks that is fine, but this helps us get a good head start on things.