Fixed Python run-time error in the Mini-Lab IPython Notebook
Fixed Python run-time error in the Mini-Lab IPython Notebook. Must be very careful of not deleting the sequence of commands to generate output.  The goal before was that each section should have the required code to generate output. It should not be dependent any other section.  This will be too hard to track down while debugging for errors.

The Python error occurred in section named "Interpreting weights".   I was able to execute the code in this section successfully without any errors.