revision 5 of lab 2
Changed from using a label encoder to using one-hot encoding using "get_dummies".  Added code to run and optimize a K Nearest Neighbors and random forest classifiers.  Added code to run a logistic regression and an SVM classifier, but still need to add code to loop through options and optimize for the best accuracy.  All classifications so far are for the income target and have not yet been duplicated for the marital status target.